+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Devop
 
 - Add sentry environment [#871](https://github.com/EthVM/EthVM/pull/871)
+- Convert all address instance to checksum [#870](https://github.com/EthVM/EthVM/pull/870)
 - Fixed spelling mistakes in readme